Reset project to FROST

+++ app/database/__init__.py
@@ -1,5 +1,6 @@
 import flask
 import flask_sqlalchemy as fsql
+import enum
 import re
 import secrets
 import sqlalchemy.dialects.mysql as sqldlc_mysql
@@ -7,21 +8,36 @@ import sqlalchemy.dialects.postgresql as sqldlc_psql
 import sqlalchemy.dialects.sqlite as sqldlc_sqlite
 import typing
 
+import app.common.utils as utils
+
 # We'll manage redis here too.
 import redis
 
-# ---------- REDIS SETUP ----------
+# ---------- REDIS Setup ----------
 redis_db: redis.StrictRedis = None
 
+
+class RedisKeyType(utils.EnumAutoName):
+    # RedisKeyType's enum values are string, because this will be used on Redis key.
+    # This is intended and must not be used on DB column type.
+    EMAIL_VERIFICATION = enum.auto()
+    EMAIL_PASSWORD_RESET = enum.auto()
+    TOKEN_REVOKE = enum.auto()
+
+    def as_redis_key(self, value: str):
+        return f'{self.value}={str(value)}'
+
+
+# ---------- RDB Setup ----------
 # Create db object when module loads, but do not connect to app context yet.
 db: fsql.SQLAlchemy = fsql.SQLAlchemy(session_options={"autoflush": False})
 
+# ---------- RDB Type Handler ----------
 PrimaryKeyType = db.BigInteger().with_variant(sqldlc_psql.BIGINT(), 'postgresql')
 PrimaryKeyType = db.BigInteger().with_variant(sqldlc_mysql.BIGINT(), 'mysql')
 PrimaryKeyType = db.BigInteger().with_variant(sqldlc_sqlite.INTEGER(), 'sqlite')
 
-
-# ---------- IntegrityError Case Checker ---------
+# ---------- RDB IntegrityError Case Checker ---------
 IntegrityCase = [
     'FAILED_NOT_NULL',
     'FAILED_UNIQUE',
@@ -33,9 +49,7 @@ IntegrityCase = [
 
 def IntegrityCaser_sqlite(err_str):
     def default_column_extractor(errstr):
-        errstr = errstr.split(':')[1]
-        errstr = errstr.split(',')
-        return list(map(lambda x: x.strip(), errstr))
+        return errstr.split(':')[1].split(',')[0].split('.')[1].strip()
 
     case_data = {
         # https://github.com/sqlite/sqlite/blob/master/src/vdbe.c#L1055
@@ -57,19 +71,19 @@ def IntegrityCaser_sqlite(err_str):
 
 def IntegrityCaser_psycopg2(err, pgcode):
     def unique_column_extractor(err):
-        return [re.findall(
-                r'Key \((\S+)\)=\((\S+)\) already exists.',
-                err.__cause__.diag.message_detail)[0], ]
+        return re.findall(
+                    r'Key \((\S+)\)=\((\S+)\) already exists.',
+                    err.__cause__.diag.message_detail)[0][0]
 
     def null_column_extractor(err):
-        return [re.findall(
-                r'null value in column "(\S+)" violates not-null constraint',
-                err.__cause__.diag.message_primary), ]
+        return re.findall(
+                    r'null value in column "(\S+)" violates not-null constraint',
+                    err.__cause__.diag.message_primary)
 
     def fk_column_extractor(err):
-        return [re.findall(
-                r'Key \((\S+)\)=\((\S+)\) is not present in table "(\S+)".',
-                err.__cause__.diag.message_detail)[0], ]
+        return re.findall(
+                    r'Key \((\S+)\)=\((\S+)\) is not present in table "(\S+)".',
+                    err.__cause__.diag.message_detail)[0][0]
 
     case_data = {
         '23000': ('', lambda x: []),  # INTEGRITY_CONSTRAINT_VIOLATION @ PSQL
@@ -137,6 +151,7 @@ def init_app(app: flask.Flask):
     import app.database.user as user  # noqa
     import app.database.board as board  # noqa
     import app.database.jwt as jwt_module  # noqa
+    import app.database.project_table as project_table  # noqa
 
     if app.config.get('RESTAPI_VERSION') == 'dev':
         # Drop DB tables when on dev mode
@@ -149,7 +164,7 @@ def init_app(app: flask.Flask):
     for table in db.get_tables_for_bind():
         table.create(checkfirst=True, bind=db.engine)
 
-    if app.config.get('RESTAPI_VERSION') == 'dev':
+    if app.config.get('RESTAPI_VERSION') == 'dev' and app.config.get('DROP_ALL_REFRESH_TOKEN_ON_LOAD', True):
         # Drop some DB tables when on dev mode
         # db.drop_all()
         db.session.query(jwt_module.RefreshToken).delete()
